ICRoom/requirements.txt
flask
redis
flask-restplus
py-bcrypt
PyJWT

